pyproject.toml
[tool.poetry]
name = "bwb-scheduler"
version = "0.1.0"
description = ""
authors = ["Varun Mittal <varunmittal91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
temporalio = "^1.9.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
networkx = "^3.2.1"
docker = "^6.1.3"
msgpack = "^1.0.7"
typer = "^0.9.0"
redis = "^5.0.1"
fastapi = "^0.110.0"
starlette-context = "^0.3.6"
uvicorn = "^0.29.0"
fsspec = "^2024.9.0"
s3fs = "^2024.9.0"
python-dotenv = "^1.0.1"
toposort = "^1.10"
paramiko = "^3.5.0"
sqlalchemy = "^2.0.37"
psutil = "^6.1.1"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.10.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start-service = "bwb_scheduler.webservice.web_app:start_app"
start-worker = "bwb_scheduler.worker:cli"

[tool.isort]
profile = "pycharm"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
lines_after_imports = -1

[tool.black]
line_length = 120
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
